removed "broker" and "resolver" and their respective components entirely; systemge has received new configs; -tcpServer: each systemge component now has a tcpServer -tcpEndpoints: each systemge component can have multiple tcpEndpoints to which it will connect;
on startup, the node will connect to the provided tcpEndpoints as client. the server will notify the client which topics it is interested in.
when sending a message, the node will propagate the message to all provided endpoints, that are interested in the given topic.

effective changes for developers:
signicantly less configs; sync requests now can receive multiple responses by different receiving nodes;
way less overhead;

commands and additional new features will be added in the days to come; spawner might be buggy and will be fixed very soon if that is the case