refactoring; reworked configs "HandleMessagesSequentially" and "ProcessMessagesSequentially"
now:
"ProcessMessagesOfEachConnectionSequentially" will guarantee, that all messages from a particular connection will be handled in the correct order.
if you only have one incoming connection this is all you need.

"ProcessAllMessagesSequentially" even if there are >1 incoming connections, this will guarantee that only one message handler at a time is called

"ProcessAllMessagesSequentiallyChannelSize" if 0, no order of arrival is guaranteed; if >0, the order of arrival is guaranteed, as long as the capacity is never reached ( will print error log if the capacity is >0 and reached)